requirements.txt
PyYAML>=6.0
simple-term-menu>=1.6.1
